Добивочка
Шлифовочка

+++ LAB5/src/Comms/Remove.java
@@ -4,7 +4,12 @@ import GivenClasses.*;
 public class Remove implements Commands{
 	public void remove_by_id(DAO<Worker> dao, long id) {
 			Worker w = dao.get(id);
-			dao.delete(w);
+			if(w == null) {
+				System.out.println("There's no guy with such id");
+			}
+			else {
+				dao.delete(w);
+			}
 	}
 	
 	@Override
